requirements.txt
boto3==1.38.12
langchain==0.3.24
langchain-aws==0.2.22
langchain-core==0.3.56
lorem==0.1.1
python-dotenv==1.1.0
streamlit==1.44.1
streamlit-local-storage==0.0.25
langgraph==0.4.1